+++ src/Braintree/IMerchantAccountGateway.cs
@@ -8,5 +8,6 @@ namespace Braintree
         Result<MerchantAccount> CreateForCurrency(MerchantAccountCreateForCurrencyRequest request);
         MerchantAccount Find(string id);
         Result<MerchantAccount> Update(string id, MerchantAccountRequest request);
+        PaginatedCollection<MerchantAccount> All();
     }
 }