+++ test/Braintree.TestUtil/TestHelper.cs
@@ -56,7 +56,7 @@ namespace Braintree.TestUtil
                 ""account_holder_name"": ""Dan Schulman"",
                 ""account_description"": ""PayPal Checking - 1234"",
                 ""ach_mandate"": {
-                    ""text"": """"
+                    ""text"": ""cl mandate text""
                 }
             }";