+++ test/Braintree.Tests.Integration/CustomerIntegrationTest.cs
@@ -134,7 +134,6 @@ namespace Braintree.Tests.Integration
             Assert.AreEqual("1234", usBankAccount.Last4);
             Assert.AreEqual("checking", usBankAccount.AccountType);
             Assert.AreEqual("Dan Schulman", usBankAccount.AccountHolderName);
-            Assert.AreEqual("PayPal Checking - 1234", usBankAccount.AccountDescription);
             Assert.IsTrue(Regex.IsMatch(usBankAccount.BankName, ".*CHASE.*"));
         }
 
@@ -617,7 +616,6 @@ namespace Braintree.Tests.Integration
             Assert.AreEqual("1234", usBankAccount.Last4);
             Assert.AreEqual("checking", usBankAccount.AccountType);
             Assert.AreEqual("Dan Schulman", usBankAccount.AccountHolderName);
-            Assert.AreEqual("PayPal Checking - 1234", usBankAccount.AccountDescription);
             Assert.IsTrue(Regex.IsMatch(usBankAccount.BankName, ".*CHASE.*"));
         }