+++ test/Braintree.Tests/TestTransactionTest.cs
@@ -5,20 +5,6 @@ namespace Braintree.Tests
     [TestFixture()]
     public class TestTransactionTest {
 
-        private BraintreeGateway gateway;
-
-        [SetUp()]
-        public void Setup()
-        {
-            gateway = new BraintreeGateway
-            {
-                Environment = Environment.DEVELOPMENT,
-                MerchantId = "integration_merchant_id",
-                PublicKey = "integration_public_key",
-                PrivateKey = "integration_private_key"
-            };
-        }
-
         [Test]
         public void FailsInProduction()
         {